ndctl, scrub: add {wait,start}-scrub helper utilities
Provide utilities to make it easy for a platform owner to poll for the
completion of ARS scrubbing, or otherwise launch an ARS run across 1 or
more nvdimm buses.

These commands take 1 or more bus identifiers (bus device name, bus
number, or bus provider name), and runs a 'start' or 'wait' ARS
operation on each.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>